Extract search highlighting related tests
